+++ test/utils/FunctionsTest.ts
@@ -144,12 +144,13 @@ suite("Functions", function FunctionsTest() {
 
 
     test("partial-many", function partialManyArgTest() {
-        var funcM1 = Functions.partial(function () { return Array.prototype.reduce.call(arguments, (s: number, i: number) => s + i, 0); });
+        var reduce = <(cb: (prev: number, cur: number, idx: number, ary: number[]) => number, defaultVal: number) => number>Array.prototype.reduce;
+        var funcM1 = Functions.partial(function () { return reduce.call(<number[]><any>arguments, (s: number, i: number) => s + i, 0); });
         asr.equal((<any>funcM1)(2, 3, 5), 10);
         asr.equal((<any>funcM1)(), 0);
         asr.equal(funcM1.name, "partialManyBindNone");
 
-        var funcM2 = Functions.partial(function () { return Array.prototype.reduce.call(arguments, (s: number, i: number) => s + i, 0); }, 1, 3);
+        var funcM2 = Functions.partial(function () { return reduce.call(<number[]><any>arguments, (s: number, i: number) => s + i, 0); }, 1, 3);
         asr.equal((<any>funcM2)(5, 9), 18);
         asr.equal((<any>funcM2)(), 4);
         asr.equal(funcM2.name, "partialManyBindMany");